Huge system changes

+++ include/vector2.h
@@ -3,7 +3,7 @@
 struct Vector2 {
     public:
         Vector2() : x(0), y(0) {}
-        Vector2(int x, int y) : x(x), y(y) {}
-        int x;
-        int y;
+        Vector2(float x, float y) : x(x), y(y) {}
+        float x;
+        float y;
 };
